django_cache/requirements.txt
django_redis
Django
djangorestframework
django-template-model